server/requirements.txt
# Core libraries
python-dotenv>=1.0.1
pydantic==2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Database and storage
redis[hiredis]==5.0.1
hvac==2.1.0

# Communication
grpcio>=1.60.0
grpcio-tools>=1.60.0
httpx[http2]>=0.25.2

# Data handling
PyJWT>=2.8.0
aiofiles>=24.1.0

# Security and Validation
bleach>=6.1.0
validators>=0.35.0
markupsafe>=2.1.0

# Logging and Monitoring
loguru>=0.7.0
prometheus-client>=0.19.0

# Token counting
tiktoken>=0.5.0

# LLM Provider SDKs (for Web Server)
openai>=1.19.1
anthropic>=0.20.0
together>=0.2.11
google-genai>=1.13.0

# HTTP client
requests>=2.32.2

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.14.0
pytest-cov>=4.1.0

# Production deployment
python-multipart>=0.0.6  # For file uploads
email-validator>=2.1.0  # For email validation
orjson>=3.9.10  # Fast JSON serialization
